Down to two entry points
