+++ package.json
@@ -64,6 +64,7 @@
     "@types/html-webpack-plugin": "^3.2.0",
     "@types/jest": "^23.3.2",
     "@types/lodash": "^4.14.115",
+    "@types/mini-css-extract-plugin": "^0.2.0",
     "@types/node": "^8.10.0",
     "@types/promise": "^7.1.30",
     "@types/react": "^16.4.14",
@@ -78,6 +79,7 @@
     "@types/sw-precache-webpack-plugin": "^0.11.0",
     "@types/uglifyjs-webpack-plugin": "^1.1.0",
     "@types/webpack": "^3.8.16",
+    "@types/webpack-config-utils": "^2.3.0",
     "@types/webpack-dev-server": "^3.1.1",
     "@types/webpack-manifest-plugin": "^1.3.2",
     "@types/webpack-node-externals": "^1.6.3",
@@ -101,6 +103,7 @@
     "husky": "1.0.1",
     "jest": "23.6.0",
     "lint-staged": "7.3.0",
+    "mini-css-extract-plugin": "^0.4.4",
     "moment": "2.22.2",
     "node-sass": "^4.9.4",
     "npm-run-all": "4.1.3",
@@ -128,6 +131,7 @@
     "uglifyjs-webpack-plugin": "1.3.0",
     "url-loader": "0.6.2",
     "webpack": "3.12.0",
+    "webpack-config-utils": "^2.3.1",
     "webpack-dev-server": "2.9.7",
     "webpack-manifest-plugin": "1.3.2",
     "webpack-node-externals": "1.7.2"
