--- scripts/app/configs/webpack.config.server.production.ts
@@ -1,81 +1,0 @@
-import ForkTsCheckerWebpackPlugin from "fork-ts-checker-webpack-plugin";
-import InterpolateHtmlPlugin from "react-dev-utils/InterpolateHtmlPlugin";
-import webpack, {Module} from "webpack";
-import webpackNodeExternals from "webpack-node-externals";
-import Application from "../Application";
-import Env from "./env";
-import files from "./files";
-import resolve from "./webpack/common/resolve";
-import {
-    typeScriptRule,
-} from "./webpack/common/rules";
-
-/**
- * Stringify all values so we can feed into Webpack DefinePlugin
- * @type Object
- */
-const env = Env.config();
-const definePluginSettings = {
-    "process.env": Object.keys(env).reduce((values, key) => {
-            values[key] = JSON.stringify(env[key]);
-            return values;
-        }, {},
-    ),
-};
-
-/**
- * Assert this just to be safe.
- */
-if (Env.current !== "production") {
-    throw new Error("Production builds must have NODE_ENV=production.");
-}
-
-const module: Module = {
-    rules: [typeScriptRule],
-    strictExportPresence: true,
-};
-
-const serverConfig = {
-    bail: true,
-    devtool: Application.sourceMapType,
-    entry: [Application.serverEntryPoint],
-    externals: [webpackNodeExternals()],
-    module,
-    output: {
-        filename: files.outputPattern.jsServer,
-        path: Application.buildPath,
-        publicPath: Application.servedPath,
-    },
-    plugins: [
-        // Makes some environment variables available in index.html.
-        // The public URL is available as %PUBLIC_URL% in index.html, e.g.:
-        // <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
-        // In production, it will be an empty string unless you specify "homepage"
-        // in `package.json`, in which case it will be the pathname of that URL.
-        new InterpolateHtmlPlugin(env.raw),
-        // Makes some environment variables available to the JS code, for example:
-        // if (process.env.NODE_ENV === 'production') { ... }. See `./env.js`.
-        // It is absolutely essential that NODE_ENV was set to production here.
-        // Otherwise React will be compiled in the very slow development mode.
-        new webpack.DefinePlugin(definePluginSettings),
-        new webpack.DefinePlugin({
-            __isBrowser__: "false",
-        }),
-        // Moment.js is an extremely popular library that bundles large locale files
-        // by default due to how Webpack interprets its code. This is a practical
-        // solution that requires the user to opt into importing specific locales.
-        // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
-        // You can remove this if you don't use Moment.js:
-        new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
-        // Perform type checking and linting in a separate process to speed up compilation
-        new ForkTsCheckerWebpackPlugin({
-            async: false,
-            tsconfig: Application.tsConfig,
-            tslint: Application.tsLint,
-        }),
-    ],
-    resolve,
-    target: "node",
-};
-
-export default serverConfig;
