+++ scripts/app/configs/webpack/common/rules.ts
@@ -20,7 +20,7 @@ const postCssPlugin = () => [
 // In production, they would get copied to the `build` folder.
 // This loader doesn"t use a "test" so it will catch all modules
 // that fall through the other loaders.
-export const catchAllRule: Rule = {
+export const catchAllRule = {
     // Exclude `js` files to keep "css" loader working as it injects
     // its runtime that would otherwise processed through "file" loader.
     // Also exclude `html` and `json` extensions so they get processed
