+++ scripts/app/configs/webpack/production/server/module.ts
@@ -1,0 +1,8 @@
+import {typeScriptRule} from "../../common/rules";
+
+const module = {
+    rules: [typeScriptRule],
+    strictExportPresence: true,
+};
+
+export default module;
